+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix the algorithm for marking all users as checked/unchecked. Refs UIPFU-115. 
 * Refactor user filtering logic in UserSearchContainer. Fixes UIPFU-107.
 * Add a new filter configuration to display the filter by user type. Refs UIPFU-99.
+* Call afterClose only after the modal state has been updated to closed. Fixes UIPFU-117.
 
 ## [8.0.0] (https://github.com/folio-org/ui-plugin-find-user/tree/v8.0.0) (2025-03-13)
 [Full Changelog](https://github.com/folio-org/ui-plugin-find-user/compare/v7.2.0...v8.0.0)